Game difficulty and Scoring Update
MAde it so you can select a game difficulty (right now starts on easy and then after 5 questions it asks if want to continue or change difficulty). When you get an answer wrong, it will continue rather than wait for you to get it right. Don't feel like adding comments at the moment, but I promise I will on some of the stuff I added